Gather compute node URL's from traefik
